Add vTPM basic tests
This change adds a test script for vtpm functionalities on XCP-ng 8.3.
The script needs a Debian based UEFI VM to create and attach a vtpm
device. It then starts the VM, runs a few tests from the VM (like
message signing), stops it and then destroys the attached vtpm device.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>